+++ src/host/directio.cpp
@@ -1051,9 +1051,9 @@ NTSTATUS SrvWriteConsoleOutputString(_Inout_ PCONSOLE_API_MSG m, _Inout_ PBOOL /
                     const ULONG elementCount = BufferSize;
                     const char* const pChars = reinterpret_cast<const char* const>(Buffer);
                     std::vector<char> chars{ pChars, pChars + elementCount };
-                    a->NumRecords = WriteOutputStringA(pScreenInfo->GetActiveBuffer(),
-                                                       chars,
-                                                       a->WriteCoord);
+                    a->NumRecords = gsl::narrow<ULONG>(WriteOutputStringA(pScreenInfo->GetActiveBuffer(),
+                                                                          chars,
+                                                                          a->WriteCoord));
                 }
                 catch (...)
                 {
@@ -1068,9 +1068,9 @@ NTSTATUS SrvWriteConsoleOutputString(_Inout_ PCONSOLE_API_MSG m, _Inout_ PBOOL /
                     const ULONG elementCount = BufferSize / sizeof(WORD);
                     const WORD* const pAttrs = reinterpret_cast<const WORD* const>(Buffer);
                     std::vector<WORD> attrs{ pAttrs, pAttrs + elementCount };
-                    a->NumRecords = WriteOutputAttributes(pScreenInfo->GetActiveBuffer(),
-                                                          attrs,
-                                                          a->WriteCoord);
+                    a->NumRecords = gsl::narrow<ULONG>(WriteOutputAttributes(pScreenInfo->GetActiveBuffer(),
+                                                                             attrs,
+                                                                             a->WriteCoord));
                     Status = STATUS_SUCCESS;
                 }
                 catch (...)
@@ -1087,9 +1087,9 @@ NTSTATUS SrvWriteConsoleOutputString(_Inout_ PCONSOLE_API_MSG m, _Inout_ PBOOL /
                     const ULONG elementCount = BufferSize / sizeof(wchar_t);
                     const wchar_t* const pChars = reinterpret_cast<const wchar_t* const>(Buffer);
                     std::vector<wchar_t> chars{ pChars, pChars + elementCount };
-                    a->NumRecords = WriteOutputStringW(pScreenInfo->GetActiveBuffer(),
-                                                       chars,
-                                                       a->WriteCoord);
+                    a->NumRecords = gsl::narrow<ULONG>(WriteOutputStringW(pScreenInfo->GetActiveBuffer(),
+                                                                          chars,
+                                                                          a->WriteCoord));
                     Status = STATUS_SUCCESS;
                 }
                 catch (...)
